requirements.txt
azure-cognitiveservices-speech==1.19.0
azure-core==1.21.1
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.0
flake8==4.0.1
Flask==2.0.2
idna==3.3
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
msrest==0.6.21
oauthlib==3.1.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
